Improving kyc whitelist script (for testing purposes)
